Change index format to have dependable size

+++ experiments/2024-10-30/src/mesh.rs
@@ -1,4 +1,4 @@
 pub struct Mesh {
     pub vertices: Vec<[f64; 3]>,
-    pub triangles: Vec<[usize; 3]>,
+    pub triangles: Vec<[u32; 3]>,
 }
